SpreadsheetContextShared extracted from SpreadsheetContextFixedSpreadsheetId
